pply retention rules during manual move of book from staging to prod
