+++ backend/tests/mill/processors/test_title.py
@@ -8,9 +8,9 @@ import pytest
 from sqlalchemy.orm import Session as OrmSession
 
 from cms_backend.db.models import Book, BookLocation, Title
+from cms_backend.db.title import sort_books_by_filename_period
 from cms_backend.mill.processors.title import (
     apply_retention_rules,
-    sort_books_by_filename_period,
 )
 from cms_backend.utils.datetime import getnow
 
@@ -177,7 +177,7 @@ def test_apply_retention_rules_keeps_last_version_of_two_most_recent_months(
     dbsession.flush()
 
     with patch(
-        "cms_backend.mill.processors.title.getnow",
+        "cms_backend.db.title.getnow",
         return_value=datetime.datetime(2024, 4, 1),
     ):
         apply_retention_rules(dbsession, title)
@@ -265,7 +265,7 @@ def test_apply_retention_rules_handles_different_flavours_separately(
     dbsession.flush()
 
     with patch(
-        "cms_backend.mill.processors.title.getnow",
+        "cms_backend.db.title.getnow",
         return_value=datetime.datetime(2024, 2, 1),
     ):
         apply_retention_rules(dbsession, title)
